🚀 Add Deletion Policy to Module Controller  (#582)
This PR introduces a new 'spec.deletionPolicy' field that defines how the Kubernetes operator handles module deletion. When set to 'retain', the module and its resources are preserved even if the custom resource is deleted. When set to 'destroy', the operator performs a destroy operation to remove all associated resources and the module itself.
---------

Co-authored-by: Aleksandr Rybolovlev <sacha.rybolovlev@hashicorp.com>